+++ tests/testthat/test-runs-bernoulli-binomial.R
@@ -22,10 +22,7 @@ bad_binomial = list(
   list(y | trials(N) ~ 1 + sigma(1)),
 
   # Weights not implemented yet
-  list(y | trials(N) + weights(weights_ok) ~ 1),
-
-  # Generalized autoregression is not implemented yet
-  list(y | trials(N) ~ 1 + ar(1))
+  list(y | trials(N) + weights(weights_ok) ~ 1)
 )
 
 test_bad(bad_binomial,
@@ -39,7 +36,8 @@ good_binomial = list(
        y | trials(N) ~ 1 ~ 1 + x,
        ~ 0),
   list(y | trials(N) ~ 1,  # With varying
-       1 + (1|id) ~ 1)
+       1 + (1|id) ~ 1),
+  list(y | trials(N) ~ 1 + ar(1))
   #list(y | trials(N) ~ 1,
   #     1 ~ N)  # N can be both trials and slope. TO DO: Fails in this test because par_x = "x"
 )
@@ -71,7 +69,7 @@ bad_bernoulli = list(
   # Weights not implemented yet
   list(y | trials(N) + weights(weights_ok) ~ 1),
 
-  # Generalized autoregression is not implemented yet
+  # Bernoulli GARMA is not currently supported
   list(y_bern ~ 1 + ar(1))
 )