+++ tests/testthat/test-runs-negbinomial.R
@@ -2,7 +2,6 @@ bad_negbinomial = list(
   list(y | trials(N) ~ 1),
   list(y ~ 1 + sigma(1)),
   list(y | weights(weights_ok) ~ 1),
-  list(y ~ 1 + ar(1)),
   list(y_bad_numeric ~ 1)
 )
 
@@ -23,7 +22,8 @@ good_negbinomial = list(
   list(
     y ~ 1 + x + shape(1),
     ~ 0 + x + shape(1)
-  )
+  ),
+  list(y ~ 1 + ar(1))
 )
 
 test_good(